# Add some basic `@wry/equality` performance tests