Typical results on my machine:
```
  performance
    ✔ should be fast for arrays (2034ms)
    ✔ should be fast for objects (1146ms)
    ✔ should be fast for strings (537ms)
    ✔ should be fast for functions (1079ms)
```
I'm currently doing some work that should noticeably improve these numbers.